% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_choropleth.R
\name{cdc_choropleth}
\alias{cdc_choropleth}
\alias{cdc_morantest}
\alias{cdc_choropleth_facet}
\title{Crear mapas coropleticos}
\usage{
cdc_choropleth(data, value, departamento, min_n, shape, title_legend,
  rute_name)

cdc_morantest(data, value, departamento)

cdc_choropleth_facet(data, value, facet, departamento, min_n, shape,
  title_legend, rute_name)
}
\arguments{
\item{data}{data frame con conteo por area administrativa}

\item{value}{nombre de columna con el valor a mapear}

\item{departamento}{departamento (adm03) a elegir dentro del pais (adm01)}

\item{min_n}{número mínimo de eventos para idnetificar polígono con su nombre}

\item{shape}{archivo sf con los datos del polígono al nivel del area administrativa}

\item{title_legend}{Título en la legenda}

\item{rute_name}{ruta y nombre de la imagen a generar}

\item{facet}{varaible name to make the facet by}
}
\value{
Mapa e imagen con conteo o tasas por área administrativa.
}
\description{
API para la generación de mapas coropleticos con información agregada por area administrativa (distrito).
}
\section{Functions}{
\itemize{
\item \code{cdc_choropleth}: 

\item \code{cdc_morantest}: moran test

\item \code{cdc_choropleth_facet}: moran test
}}

\examples{

# # librerias ---------------------------------------------------------------
#
# library(tidyverse)
# library(readxl)
# library(tmap)
# library(sf)
#
# #casos ---------------------------------------------------------------
# caso_per <- read_xls("data-raw/ubigeo_casos.xls") \%>\%
#   mutate(n_depa=str_to_title(n_depa))
#
# #shape ---------------------------------------------------------------
#
# per0 <- st_read("data-raw/gis/PER_adm0.shp")
# per1 <- st_read("data-raw/gis/PER_adm1.shp")
# per2 <- st_read("data-raw/gis/PER_adm2.shp")
# per4 <- st_read("data-raw/gis/Distritos.shp")
#
# #ejemplo ---------------------------------------------------------------
# cdc_choropleth(data = edas_aggr,
#               value = rate,
#               departamento = "Cajamarca",
#               min_n = 5,
#               shape = per4,
#               title_legend = "Tasa cruda\\nde EDA\\nx 1000 hab.",
#               rute_name = "figure/fig22-eda_fun-")
#
# cdc_choropleth(data = edas_aggr,
#               value = casos,
#               departamento = "Cajamarca",
#               min_n = 200,
#               shape = per4,
#               title_legend = "Casos\\nde EDA",
#               rute_name = "figure/fig23-eda_fun-")
#
# cdc_choropleth_facet(data = edas_aggr_semanal,
#                      value = casos_cumsum,
#                      facet = semana,
#                      departamento = "Piura",
#                      min_n = 5,
#                      shape = per4,title_legend = "Casos\\nacumulados\\nde EDA",
#                      rute_name = "figure/fig27-eda_fun-")
#
# cdc_morantest(data = edas_aggr,value = rate,departamento = "Lambayeque")
# cdc_morantest(data = edas_aggr,value = casos,departamento = "Lambayeque")

}
