% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_coropleth.R
\name{cdc_coropleth}
\alias{cdc_coropleth}
\title{Crear mapas coropleticos}
\usage{
cdc_coropleth(data, departamento, min_n, title_legend, shape, rute_name)
}
\arguments{
\item{data}{data frame con conteo por area administrativa}

\item{departamento}{departamento (adm03) a elegir dentro del pais (adm01)}

\item{min_n}{número mínimo de eventos para idnetificar polígono con su nombre}

\item{title_legend}{Título en la legenda}

\item{shape}{archivo sf con los datos del polígono al nivel del area administrativa}

\item{rute_name}{ruta y nombre de la imagen a generar}
}
\value{
Mapa e imagen con conteo o tasas por área administrativa.
}
\description{
API para la generación de mapas coropleticos con información agregada por area administrativa (distrito).
}
\examples{

# librerias ---------------------------------------------------------------

library(tidyverse)
library(readxl)
library(tmap)
library(sf)

#casos ---------------------------------------------------------------
caso_per <- read_xls("data-raw/ubigeo_casos.xls") \%>\%
  mutate(n_depa=str_to_title(n_depa))

#shape ---------------------------------------------------------------

per0 <- st_read("data-raw/gis/PER_adm0.shp")
per1 <- st_read("data-raw/gis/PER_adm1.shp")
per2 <- st_read("data-raw/gis/PER_adm2.shp")
per4 <- st_read("data-raw/gis/Distritos.shp")

#ejemplo ---------------------------------------------------------------
cdc_coropleth(data = caso_per,
              departamento = "Junin",
              min_n = 2,
              title_legend = "Casos\\nde SGB",
              shape = per4,
              rute_name = "figure/fig04-sbg_fun-")


# loop --------------------------------------------------------------------

d_list <- c("Lima","Junin","La Libertad","Cajamarca","Lambayeque","Piura")

for (i in 1:length(d_list)) {
  cdc_coropleth(data = caso_per,
                departamento = d_list[i],
                min_n = 2,
                title_legend = "Casos\\nde SGB",
                shape = per4,
                rute_name = "figure/fig04-sbg_fun-")
}

}
