% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_misc.R
\name{cdc_datatable_html}
\alias{cdc_datatable_html}
\title{Unclassified functions}
\usage{
cdc_datatable_html(data)
}
\arguments{
\item{data}{input dataset}
}
\value{
html table
}
\description{
Create html table
}
\section{Functions}{
\itemize{
\item \code{cdc_datatable_html}: 
}}

\examples{

library(tidyverse)
iris \%>\% group_by(Species) \%>\% summarise_all(.funs = mean) \%>\% cdc_datatable_html()

}
