% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_special.R
\name{cdc_dotwhiskers_plot}
\alias{cdc_dotwhiskers_plot}
\alias{cdc_dotwhiskers_plot_2}
\title{Make some special plots}
\usage{
cdc_dotwhiskers_plot(data, var_categorical, var_continuous)

cdc_dotwhiskers_plot_2(data, var_continuous, ...)
}
\arguments{
\item{data}{base de datos}

\item{var_categorical}{variable categórica (idealmente politómica)}

\item{var_continuous}{variable numérica contínua}

\item{...}{add multiple categorical variables}
}
\value{
estimacion puntual e incertidumbre de la distribución de una variable por cada nivel de categorias
}
\description{
Make a dot-whiskers plot (a.k.a starwars TE fighter spaceship plot!)
}
\section{Functions}{
\itemize{
\item \code{cdc_dotwhiskers_plot}: 

\item \code{cdc_dotwhiskers_plot_2}: priorización con dos covariables
}}

\examples{

# inspiration: https://twitter.com/robinson_es/status/1193204992120958976/photo/1

library(tidyverse)
library(charlatan)

n_obs <- 33
set.seed(n_obs)

ch_data_wide <- tibble(
  #values
  values_01 = ch_integer(n = n_obs,min = 5,max = 10),
  values_02 = ch_integer(n = n_obs,min = 5,max = 20),
  values_03 = ch_integer(n = n_obs,min = 5,max = 30),
  category = ch_integer(n = n_obs,min = 0,max = 1) \%>\% as.logical()) \%>\%
  pivot_longer(cols = values_01:values_03,
               names_to = "variable",
               values_to = "value")

# ch_data_wide \%>\%
#   group_by(variable) \%>\%
#   skimr::skim_without_charts()

cdcper::cdc_dotwhiskers_plot(data = ch_data_wide,
                             var_categorical = variable,
                             var_continuous = value) \%>\%
  ggplot(aes(x = estimate,y = variable)) +
  geom_errorbarh(aes(xmax = conf.high, xmin = conf.low)) +
  geom_point()

cdcper::cdc_dotwhiskers_plot_2(data = ch_data_wide,
                               var_continuous = value,
                               category, variable)


}
