% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_egresos.R
\name{cdc_egresos_summary}
\alias{cdc_egresos_summary}
\alias{cdc_egresos_by_year}
\alias{cdc_piramide}
\title{Crear bases y tablas de conteo yfrecuencia de egresos hospitalarios}
\usage{
cdc_egresos_summary(data, ..., percent_by = c(1), save = FALSE,
  rute = "data/")

cdc_egresos_by_year(data_long, rute = "table/")

cdc_piramide(data, x = pct, y = e_quinq_2, z = sexo)
}
\arguments{
\item{data}{data frame con información nominal (por sujeto)}

\item{...}{lista de una o más variables}

\item{percent_by}{(numeric vector) seleccionar las variables para generar porcentajes}

\item{save}{(logic) opción para grabar o no la base resumen generada}

\item{rute}{(character string) ruta del objeto a exportar. Las bases se exportan en RDS y DTA. Las tablas en CSV y XLSX.}

\item{data_long}{output from cdc_egresos_summary}

\item{x}{variable en eje x de la pirámide}

\item{y}{variable en eje y de la pirámide}

\item{z}{variable en eje z de la pirámide}
}
\value{
Base con conteo y porcentajes por cruce de variables
}
\description{
Generación de bases resumen para la creación de gráficos de tendencia y tablas reporte con información nominal (por sujeto).
}
\section{Functions}{
\itemize{
\item \code{cdc_egresos_summary}: 

\item \code{cdc_egresos_by_year}: create denominator

\item \code{cdc_piramide}: create denominator
}}

\examples{
egress <- data(egresos)
#explore data
egress \%>\% glimpse()
#in one line
cdc_egresos_summary(data = egress,year,nombdep)
cdc_egresos_summary(data = egress,percent_by = c(1,2),year,nombdep,sexo)
#cdc_egresos_summary(data = egress,percent_by = c(2,3),year,nombdep,sexo)
cdc_egresos_summary(data = egress,year,nombdep,edad_asis2015,sexo)
cdc_egresos_summary(data = egress,percent_by = c(1,2,3),year,nombdep,edad_asis2015,sexo)
cdc_egresos_summary(data = egress,year,desc_12)
cdc_egresos_summary(data = egress,percent_by = c(1,2),year,nombdep,desc_12)
#with tidyverse
egress \%>\%
  filter(year==max(year)) \%>\%
  cdc_egresos_summary(year,desc_12) \%>\%
  arrange(desc(pct))
#output
egress \%>\%
  cdc_egresos_summary(year,nombdep,save = T) #dataset
egress \%>\%
  cdc_egresos_summary(percent_by = c(1,2),
                      year,nombdep,edad_asis2015) \%>\%
  cdc_egresos_by_year() #wide table
#plot piramide
egress \%>\%
  filter(year==max(year)) \%>\%
  cdc_egresos_summary(year,e_quinq_2,sexo) \%>\%
  cdc_piramide() #class ggplot

}
