% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_channel.R
\name{cdc_endemic_channel_mutate}
\alias{cdc_endemic_channel_mutate}
\alias{cdc_endemic_channel_join}
\alias{cdc_endemic_channel_ggplot}
\title{Crear canal endémico}
\usage{
cdc_endemic_channel_mutate(disease, population, name,
  method = "gmean_1sd")

cdc_endemic_channel_join(disease_channel, disease_now)

cdc_endemic_channel_ggplot(joined_channel)
}
\arguments{
\item{disease}{disease surveillance datasets}

\item{population}{estimated population for each adm area}

\item{name}{free code name of diseases}

\item{method}{specify the method. "gmean_1sd" is geometric mean w/ 1 standard deviation (default). "gmean_2sd" is gmean w/ 2 sd. "gmean_ci" is gmean w/ 95 percent confidence intervals.}

\item{disease_channel}{salida de cdc_*_mutate}

\item{disease_now}{nueva base de vigilancia}

\item{joined_channel}{base de datos unida}
}
\value{
canal endemico, union y grafico
}
\description{
Generar estimaciones y detección de brotes.
}
\section{Functions}{
\itemize{
\item \code{cdc_endemic_channel_mutate}: 

\item \code{cdc_endemic_channel_join}: unir resultado con vigilancia actual

\item \code{cdc_endemic_channel_ggplot}: crear gráfico
}}

\examples{

# not yet

}
