% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_point_heatmap.R
\name{cdc_heatmap}
\alias{cdc_heatmap}
\alias{cdc_point_heatmap}
\alias{cdc_point_heat_mark}
\title{Crear mapas con patrones puntuales}
\usage{
cdc_heatmap(data_shape, data_point, rute_name = NA, subset_list, mp_xlim,
  mp_ylim, mp_dist, mp_loca)

cdc_point_heatmap(data_shape, data_point, rute_name = NA, subset_list,
  mp_xlim, mp_ylim, mp_dist, mp_loca)

cdc_point_heat_mark(data_shape, data_point, mp_xlim, mp_ylim, mp_dist,
  mp_loca)
}
\arguments{
\item{data_shape}{pending}

\item{data_point}{pending}

\item{rute_name}{pending}

\item{subset_list}{pending}

\item{mp_xlim}{pending}

\item{mp_ylim}{pending}

\item{mp_dist}{pending}

\item{mp_loca}{pending}
}
\value{
Mapa e imagen con casos distribuidos espacialmente.
}
\description{
Generar de mapas con patrones puntuales.
}
\section{Functions}{
\itemize{
\item \code{cdc_heatmap}: 

\item \code{cdc_point_heatmap}: marked heat map

\item \code{cdc_point_heat_mark}: marked heat map
}}

\examples{

# per4 <- st_read("data-raw/gis/Distritos.shp")
# juni <- st_read("data-raw/gis/Junin.shp") \%>\% filter(prov_res=="HUANCAYO")
#
# # edit polygons -------------------------------------------------------------------
#
# jun4 <- per4 \%>\% filter(NM_DEPA=="JUNIN") \%>\% filter(NM_PROV=="HUANCAYO")
#
# cdc_point_heatmap(data_shape = jun4,
#                   data_point = juni,
#                   rute_name = "figure/fig10-fun-sgb-junin") +
#   ggtitle("JUNIN - Provincia: Huancayo")
# ggsave(paste0("figure/fig10-fun-sgb-junin",".png"),height = 10,width = 10)
#
#
# # con una marca -----------------------------------------------------------
#
# cdc_point_heat_mark(data_point = piur,
#                     data_shape = piu4,
#                     mp_xlim = c(-81.4,-80.2),
#                     mp_ylim = c(-5.45,-4.4),
#                     mp_dist = 10,
#                     mp_loca = "bottomright")

}
