% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_morbilidad_map.R
\name{cdc_mapa_morb}
\alias{cdc_mapa_morb}
\title{Crear mapa coropletico}
\usage{
cdc_mapa_morb(data, prov, enf, anio, tip,
  shape_departamento = "C:/Users/USER/Dropbox/cdc_peru/shape_r/Limites/departamento_geo.shp",
  shape_provincia = "C:/Users/USER/Dropbox/cdc_peru/shape_r/Limites/provincia_geo.shp",
  shape_distrito = "C:/Users/USER/Dropbox/cdc_peru/shape_r/Limites/DISTRITOS.shp",
  poblacion_distrito_csv = "C:/Users/USER/Dropbox/cdc_peru/shape_r/Limites/pob_inei.csv")
}
\arguments{
\item{enf}{nombre de la enfermedad segun CIE-10}

\item{anio}{anio o intervalo de anios}

\item{shape_departamento}{shape_departamento}

\item{shape_provincia}{shape_provincia}

\item{shape_distrito}{shape_distrito}

\item{poblacion_distrito_csv}{poblacion_distrito_csv}

\item{provincia}{nombre de la provincia}

\item{tipo}{1=Conteo total & 2=Tasas}
}
\value{
Diagrama de Pareto por Area administrativa.
}
\description{
Generacion de mapas coropleticos por conteo y/o por tasas, segun area administrativa.
}
\section{Functions}{
\itemize{
\item \code{cdc_mapa_morb}: funcion para crear mapa coropletico
}}

\examples{

# library(tidyverse)
# library(magrittr)
# library(dplyr)
# library(sf)

# # Extraemos la base de datos
# lima<-cdc_extraer("lima")
# # Diagrama de Pareto por departamento
# cdc_mapa_morb(lima, "LIMA","Anemias nutricionales",2015,2)
# Uso de pipe
# lima\%>\%cdc_mapa_morb(provincia = "LIMA",distrito = "Villa El Salvador",year=2016)

}
