% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_mortality.R
\name{cdc_mortality}
\alias{cdc_mortality}
\alias{cdc_denominator}
\alias{cdc_mortality_2}
\alias{cdc_mortality_adj}
\alias{cdc_mortality_raw}
\alias{cdc_mortality_to_wide}
\title{Crear bases y tablas de conteo y tasas estandarizadas de mortalidad}
\usage{
cdc_mortality(data, denominador, estandar, ...,
  rute_name = NA_character_)

cdc_denominator(denominador, estrato, ...)

cdc_mortality_2(data, denominador, estandar, ...,
  rute_name = NA_character_)

cdc_mortality_adj(data, denominador, estandar, ...,
  rute_name = NA_character_)

cdc_mortality_raw(data, denominador, ..., rute_name = NA_character_)

cdc_mortality_to_wide(data_long, rute_name = NA)
}
\arguments{
\item{data}{data frame con información nominal (por sujeto)}

\item{denominador}{data frame con cantidad estimada de personas por rango de edad, sexo, departamento y año}

\item{estandar}{data frame con proporciones otorgadas por la OMS para estandarizar tasas (fuente: https://seer.cancer.gov/stdpopulations/world.who.html)}

\item{...}{lista de una o más variables}

\item{rute_name}{character string con la ruta y nombre del objeto a exportar. Las bases se exportan en RDS y DTA. Las tablas en CSV y XLSX.}

\item{estrato}{estrato de edad a usar para estandarizar}

\item{data_long}{output from cdc_mortality}
}
\value{
Base con conteo, población expuesta, tasas crudas y estandarizadas.
}
\description{
API para la generación de bases resumen para la creación de gráficos de tendencia y tablas reporte con información nominal (por sujeto).
}
\section{Functions}{
\itemize{
\item \code{cdc_mortality}: 

\item \code{cdc_denominator}: create denominator

\item \code{cdc_mortality_2}: create mortality with extra variables

\item \code{cdc_mortality_adj}: create mortality with extra variables

\item \code{cdc_mortality_raw}: create mortality with extra variables

\item \code{cdc_mortality_to_wide}: create output table
}}

\examples{

# ISSUE: subir base denominadores al paquete!

# # dependencies ------------------------------------------------------------
#
# #flexibility on dot-dot-dot
# #https://stackoverflow.com/questions/7028385/can-i-remove-an-element-in-dot-dot-dot-and-pass-it-on
# #https://www.burns-stat.com/the-three-dots-construct-in-r/
#
# library(tidyverse)
# library(magrittr)
# library(readxl)
# library(haven)
# library(naniar)
# library(ggrepel)
# library(labelled)
# library(gridExtra)
# library(xlsx)
#
# # importar base ----------------------------------------------------------------
#
# mortdb <- data(mortdb)
#
# # importar denominadores -------------------------------------------------------
#
# denomi_pe <- read_rds("data/pop_pe-year_dept_age_std_sex-2000_2020.rds") \%>\%
#   mutate(sexo=str_to_title(sexo))
#
# fctwho_std <- read_rds("data/who-age_std-2000_2025.rds") #same estandard per sex
#
# # CDR + DAR ---------------------------------------------
#
# # formato de base nominal
# mortdb \%>\% glimpse()
#
# #cuando mort y deno comparten covariables
# cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year)
# cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year,sexo)
# cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year,sexo,departamento)
# cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year,departamento)
# cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year)
#
# #cuando mort y deno no comparten covariables
# deno_ys <- cdc_denominator(denominador = denomi_pe,year,sexo)
# cdc_mortality_2(data = mortdb,
#                 denominador = deno_ys,
#                 estandar = fctwho_std,
#                 year,sexo,descrip_grupo10)
# #o todo en una
# cdc_mortality_2(data = mortdb,
#                 denominador = cdc_denominator(denominador = denomi_pe,year),
#                 estandar = fctwho_std,
#                 year,edad_q)
# cdc_mortality_2(data = mortdb \%>\% filter(edad_estandar=="0-4"),
#                 denominador = cdc_denominator(denominador = denomi_pe,year),
#                 estandar = fctwho_std,
#                 year,edad_estandar)
#
# #cuando no se requiere tasas
# mortdb \%>\% count(year,edadmef) #edad_fertil
# mortdb \%>\% count(year,edad_asis2015) #etapas de vida
#
# # opcionalmente, se puede exportar en RDS y DTA
# cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std,year,sexo,
#               rute_name = "data/mortalidad_anual_nacional_sexo")
#
# # formato wide para exportar tabla resumen en CSV y XLSX
# mortdb \%>\%
#   cdc_mortality(denominador = denomi_pe,estandar = fctwho_std,year,sexo,departamento) \%>\%
#   cdc_mortality_to_wide(rute_name = "table/mortalidad_anual_nacional_sexo_departamento")
#

}
