% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_morbilidad_plot.R
\name{cdc_pareto}
\alias{cdc_pareto}
\title{Crear diagrama de Pareto para datos de morbilidad}
\usage{
cdc_pareto(departamento, provincia = NULL, distrito = NULL, year)
}
\arguments{
\item{departamento}{nombre del departamento extraido por cdc_extraer}

\item{provincia}{nombre de la provincia, tipo caracter}

\item{distrito}{nombre del distrito, tipo caracter}

\item{year}{year o intervalo de years}
}
\value{
Diagrama de Pareto por Area administrativa.
}
\description{
GeneraciOn de diagrama de Pareto para datos de morbilidad, tablas de frecuencias y frecuecias acumuladas, por Area administrativa.
}
\section{Functions}{
\itemize{
\item \code{cdc_pareto}: funcion para crear diagrama de Pareto
}}

\examples{

# library(tidyverse)
# library(magrittr)
# library(dplyr)

# # Extraemos la base de datos
# lima<-cdc_extraer("lima")
# # Diagrama de Pareto por departamento
# cdc_pareto(lima,year=2012)
# # Diagrama de Pareto por provincia
# cdc_pareto(lima,provincia="LIMA",year=2014)
# # Diagrama de Pareto por distrito
# cdc_pareto(lima,provincia = "LIMA",distrito = "Villa El Salvador",year=2016)
# Uso de pipe
# lima\%>\%cdc_pareto(provincia = "LIMA",distrito = "Villa El Salvador",year=2016)

}
