% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_priorities.R
\name{cdc_pareto_lista}
\alias{cdc_pareto_lista}
\alias{cdc_pareto_lista_2}
\alias{cdc_carga_coalesce}
\title{Priorizar listados en base a multiples covariables}
\usage{
cdc_pareto_lista(data, variable, pareto_cut = 85)

cdc_pareto_lista_2(data, variable_c, variable_d, pareto_cut = 80)

cdc_carga_coalesce(data)
}
\arguments{
\item{data}{base de datos}

\item{variable}{variable numérica contínua bajo la cual priorizar algún listado}

\item{pareto_cut}{punto de corte tipo pareto}

\item{variable_c}{variable continua}

\item{variable_d}{variable discreta}
}
\value{
selección por criterio pareto y coalescencia para multiples covariables
}
\description{
Mostrar qué elementos representan el 80% de la suma total
}
\section{Functions}{
\itemize{
\item \code{cdc_pareto_lista}: 

\item \code{cdc_pareto_lista_2}: priorización con dos covariables

\item \code{cdc_carga_coalesce}: coalescencia de multiples covariables
}}

\examples{

# not yet

}
