% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_spclustr.R
\name{cdc_rsatscan_clean}
\alias{cdc_rsatscan_clean}
\title{Crear mapas coropleticos con resultados de SatScan}
\usage{
cdc_rsatscan_clean(rsatscan_in, rsatscan_out)
}
\arguments{
\item{rsatscan_in}{data de ingreso a rsatscan}

\item{rsatscan_out}{salida de rsatscan}
}
\value{
Limpiar salida estadística de rsatscan.
}
\description{
Limpiar salida y generar de mapas coropleticos con resultados estadísticos de R SatScan (rsatscan).
}
\section{Functions}{
\itemize{
\item \code{cdc_rsatscan_clean}: funcion para limpiar salida de satscan y generar mapas
}}

\examples{

#
# library(tidyverse)
# library(janitor)
# library(rsatscan)
# library(sf)
# library(tmap)
# library(lubridate)
# library(rlang)
# library(cdcper)
# theme_set(theme_bw()) #fondo blanco y negro del ggplot
#
# td = tempdir()
# mysatscanloc <- "C:/Program Files (x86)/SaTScan"
#
# # importar  ---------------------------------------------------------------
#
# edas_aggr <- read_rds("data/20190702-edas_ubigeo_nacional_sem15.rds")
# edas_aggr_semanal <- read_rds("data/20190702-edas_ubigeo_semana.rds")
#
# edas_aggr \%>\% glimpse()
# edas_aggr_semanal \%>\% glimpse()
#
# # 02 - weekly data --------------------------------------------------------
#
# d_list <- c("LIMA",
#             "Junin","La Libertad","Cajamarca","Lambayeque","Piura")
#
# name_dptx <- d_list[3]
#
# edas_aggr_dept <- edas_aggr \%>\%
#   filter(n_depa==name_dptx)
#   #filter(n_provincia==name_dptx)
#
# edas_aggr_semanal_dept <- edas_aggr_semanal \%>\%
#   filter(n_depa==name_dptx) \%>\%
#   #filter(n_provincia==name_dptx) \%>\%
#   filter(date>="2019-04-07") #fijar fechas desde aqui!
#
# NMcas_ <- edas_aggr_semanal_dept \%>\% select(ubigeo,casos,date_f) \%>\% as.data.frame()
# NMpop_ <- edas_aggr_semanal_dept \%>\% select(ubigeo,ano,total) \%>\% distinct() \%>\% as.data.frame()
# NMgeo_ <- edas_aggr_semanal_dept \%>\% select(ubigeo,lat_ctd,lon_ctd) \%>\% as.data.frame()
# head(NMcas_)
# head(NMpop_)
# head(NMgeo_)
#
# # analisis 02 espacial puro - datos semana ----------
# # analisis espacial puro
# # modelo poisson discreto
# # con datos por semana - mejor estimacion
#
# write.cas(NMcas_, td,"NM")
# write.geo(NMgeo_, td,"NM")
# write.pop(NMpop_, td,"NM")
#
# invisible(ss.options(reset=TRUE))
# ss.options(list(CaseFile="NM.cas",
#                 PopulationFile="NM.pop",
#                 CoordinatesFile="NM.geo",
#                 StartDate="2019/4/13",EndDate="2019/6/29", #days
#                 CoordinatesType=1,
#                 AnalysisType=1,
#                 ModelType=0,
#                 TimeAggregationUnits=3,
#                 PrecisionCaseTimes=3,
#                 NonCompactnessPenalty=0,
#                 ReportGiniClusters="n",
#                 LogRunToHistoryFile="n"
# ))
#
# write.ss.prm(td,"testnm")
# testnm = satscan(td,"testnm", sslocation=mysatscanloc)
# summary(testnm)
# summary.default(testnm)
#
# # output limpio!
# week_pure_spatial <- cdc_rsatscan_clean(rsatscan_in = edas_aggr_dept, #ojo! in dato agregado
#                                         rsatscan_out = testnm)
# week_pure_spatial

}
