% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_eda_ira.R
\name{clean_eda_sp}
\alias{clean_eda_sp}
\alias{clean_ira_sp}
\alias{cdc_casos_tiempo}
\alias{cdc_casos_nacional}
\title{Limpiar bases de EDA e IRA}
\usage{
clean_eda_sp(data, strata)

clean_ira_sp(data, strata)

cdc_casos_tiempo(data, value, denominador, rate_mult)

cdc_casos_nacional(data, value, denominador, rate_mult)
}
\arguments{
\item{data}{data frame con conteo por area administrativa}

\item{strata}{nombre de columna con el valor a estatificar el conte (establecimiento o ubigeo)}

\item{value}{nombre de columna del valor a contar (e.g. eda, dis, daa)}

\item{denominador}{base de denominador}

\item{rate_mult}{factor a escalar tasa}
}
\value{
Bases limpiar de EDA e IRA.
}
\description{
Funciones para limpiar bases de EDA e IRA.
}
\section{Functions}{
\itemize{
\item \code{clean_eda_sp}: limpiar base EDA

\item \code{clean_ira_sp}: limpiar base IRA

\item \code{cdc_casos_tiempo}: crear resumen por semana

\item \code{cdc_casos_nacional}: crear resumen agregado por año
}}

\examples{

#
# # importar casos ------------------------------------------------------------------
#
# iras_i <- read.dbf("data-raw/eda_ira/20190702-IRA_SP.dbf") \%>\% as_tibble()
# edas_i <- read.dbf("data-raw/eda_ira/20190702-EDA_SP.DBF") \%>\% as_tibble()
#
# edas_i \%>\% glimpse()
# iras_i \%>\% glimpse()
# edas \%>\% count(UBIGEO)
#
# file_date <- "20190702"
#
# # 1. limpiar bases ------------------------------------------------------------
#
# edas <- edas_i \%>\% clean_eda_sp(strata = ubigeo)
# edas \%>\% glimpse()
#
# iras <- iras_i \%>\% clean_ira_sp(strata = ubigeo)
# iras \%>\% glimpse()
#
# # tarea: homogenizar denominador -------------------------------------------------
#
# #1871
# denomi_pe
# #1855
# edas \%>\% count(ubigeo)
#
# # 2. calcular casos, acumulado, tasa por ubigeo ------------------------------------
#
# #base 1
# edas_aggr_semanal <- edas \%>\%
#   cdc_casos_tiempo(value = eda_c_sum,
#                    denominador = denomi_pe) \%>\%
#   write_rds(paste0("data/",file_date,"-edas_ubigeo_semana.rds"))
#
# iras \%>\%
#   cdc_casos_tiempo(value = ira_c_sum,
#                    denominador = denomi_pe) \%>\%
#   write_rds(paste0("data/",file_date,"-iras_ubigeo_semana.rds"))
#
# # base 2
# edas_aggr <- edas \%>\%
#   filter(semana>=15) \%>\%
#   cdc_casos_nacional(value = eda_c_sum,
#                      denominador = denomi_pe) \%>\%
#   write_rds(paste0("data/",file_date,"-edas_ubigeo_nacional_sem15.rds"))
#
# iras \%>\%
#   filter(semana>=15) \%>\%
#   cdc_casos_nacional(value = ira_c_sum,
#                      denominador = denomi_pe) \%>\%
#   write_rds(paste0("data/",file_date,"-iras_ubigeo_nacional_sem15.rds"))

}
