% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_data.R
\docType{data}
\name{denominadores_departamento_2000_2020_peru}
\alias{denominadores_departamento_2000_2020_peru}
\title{CDC databases: Department population 2000-2020}
\format{An object of class dataset as tibble.}
\source{
MINSA-REUNIS, <https://www.minsa.gob.pe/reunis/data/poblacion_estimada.asp> ; World (WHO 2000-2025) Standard <https://seer.cancer.gov/stdpopulations/world.who.html>
}
\usage{
data(denominadores_departamento_2000_2020_peru)
}
\description{
Department population 2000-2020
}
\examples{
library(cdcper)
library(tidyverse)
data(denominadores_departamento_2000_2020_peru)
denominadores_departamento_2000_2020_peru \%>\% glimpse()

}
\references{
MINSA
([REUNIS](https://www.minsa.gob.pe/reunis/data/poblacion_estimada.asp))
}
\keyword{datasets}
