% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_met_doc.R
\name{get_data_gri}
\alias{get_data_gri}
\alias{get_ids}
\alias{var_fun}
\title{Descarga de información meteorológica}
\usage{
get_data_gri(var, dep = NULL, pro = NULL, dis = NULL, ids = NULL,
  fuente = NULL, ruta = NULL, rango = NULL, esc_temp = NULL,
  down = FALSE, filename = paste0("Grilla_", var, ".csv"))

get_ids(fuente = "local", ruta = NULL)

var_fun(x, var)
}
\arguments{
\item{var}{Variable que será buscada. Puede ser "Prec", "Tmax" o "Tmin".}

\item{dep}{Nombre o código (en caracter) del departamento.}

\item{pro}{Nombre o código (en caracter) de la provincia.}

\item{dis}{Nombre o código (en caracter) del distrito.}

\item{ids}{Objeto que contiene la metadata.}

\item{fuente}{Ubicación de la base de datos. Esta puede ser "local" o "web".}

\item{ruta}{Dirección de la carpeta que contiene los datos. Solo funciona cuando fuente = "local".}

\item{rango}{Fecha inicial y final que será filtrada.}

\item{esc_temp}{Escala temporal en la que se descargará la información. Puede ser "Semana", "Mes", "Año" o NULL (vendría a ser escala diaria).}

\item{down}{Indique si la información se descargará en un archivo (TRUE) o no (FALSE).}

\item{filename}{Nombre del archivo donde se exportará la información.}

\item{x}{Datos serán sumados (precipitación) o promediados (el resto).}

\item{var}{Nombre de la variable}
}
\value{
Tabla con información de una variable meteorológica para una lugar, tiempo y escala temporal específica.
}
\description{
Funciones que permiten la descarga de información meteorológica
}
\details{
Descargo de responsabilidad y usos de la información

La información que se puede obtener con estas funciones no ha pasado por un control de calidad ni validación con mediciones en tierra. Esta información es una estimación de las condiciones reales atmosféricas en un lugar y tiempo determinado. Las condiciones reales son aquellas obtenidas a través de mediciones directas (por ejemplo, los datos de estaciones meteorológicas).
Ventajas

- Información para todos los distritos del Perú. Se entiende que se puede obtener información a nivel provincial y departamental también.

- Información continua desde el 2000 hasta el 2019. La información se actualiza a mediados de cada mes con la información del mes anterior. Por ejemplo, para mediados de febrero se tendrá información actualizada hasta enero del 2020.

- A pesar de no haber pasado por un control de calidad, esta información sirve de muy buena referencia para entender el comportamiento de la atmósfera y el clima. Además, los eventos extremos meteorológicos pueden ser encontrados en esta información.

- Las fuentes de información mencionadas son ampliamente utilizadas en el área de Meteorología y Climatología en muchos trabajos de investigación.

Desventajas

- Al ser información sin control de calidad es posible que la información tenga errores de sobrestimación o subestimación, pero el comportamiento temporal no debería presentar muchas diferencias.

- Mientras la resolución temporal sea más alta (por ejemplo: información diaria), el error de los datos también será mayor.

- La información distrital, provincial y departamental representa a toda el área geográfica y no a una ubiación espacial específica. Por ejemplo, la información del distrito de Chachapoyas es representativa de todo el terreno dentro del distrito y no de la ubiación de la ciudad de Chachapoyas.

- No es posible hablar de cambio climático con los casi 20 años de información que se tienen, pero es posible describir el comportamiento climático normal.

Usos

- La información debe ser utilizada de manera referencial y no debe ser considerado como el dato real.

- Utilizar la información para asociar el comportamiento temporal del clima con alguna variable a estudiar.

- La información debe ser utilizada como representativa del área de un distrito y no de una ubicación específica.

- La información puede ser utilizada para determinar el comportamiento espacial del clima entre distritos

- La información puede ser utilizada para analizar eventos extremos, mas no para determinar el valor numérico del evento.
}
\section{Functions}{
\itemize{
\item \code{get_data_gri}: función que realiza la descarga de información

\item \code{get_ids}: función que contiene la ubicación de la información

\item \code{var_fun}: función para realizar suma o promedio dependiendo de la variable asignada
}}

\examples{

library(dplyr)
library(googledrive)

# Por primera vez: poner 1 -> ir a correo -> aceptar
ids.web <- get_ids(fuente = "web")

prec <- get_data_gri(
   var = "Prec",
	  dep = "LORETO",
   pro = "MAYNAS",
   dis = "IQUITOS",
   fuente = "web",
   ids = ids.web,
   esc_temp = "Mes",
   rango = c("2010-01-01","2018-12-31"))

# final product
prec

}
