% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_forecast.R
\name{gg_forecast}
\alias{gg_forecast}
\title{Forecast plot with sw_seep() output}
\usage{
gg_forecast(data, axis_x = index)
}
\arguments{
\item{data}{input dataset}

\item{axis_x}{variable with time axis type "yearmon"}
}
\value{
ggplot
}
\description{
Create ggplot of forecast plot with sw_seep() output
}
\section{Functions}{
\itemize{
\item \code{gg_forecast}: 
}}

\examples{

# from: ?sw_sweep

library(forecast)
library(sweep)
library(dplyr)
library(ggplot2)

# ETS forecasts
a <- USAccDeaths \%>\%
  ets() \%>\%
  forecast(level = c(80, 95, 99)) \%>\%
  sw_sweep()

# crear false observed values for predicted time units
b <- a \%>\%
  filter(key=="forecast") \%>\%
  mutate(key="observed") \%>\%
  mutate(value=mean(value)) \%>\%
  mutate(across(.cols = c(lo.80:hi.99),
         .fns = ~(.x=NA_real_)))

# create plot
a \%>\%
  union_all(b) \%>\%
  # avallecam::print_inf()
  gg_forecast()

}
