% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_morbilidad_data.R
\name{his_extraer}
\alias{his_extraer}
\title{Llamar base de datos de morbilidad consolidada segun Area administrativa.}
\usage{
his_extraer(departamento, provincia = NULL, distrito = NULL,
  ruta = "C:/Users/USER/Dropbox/cdc_peru/his_r/data/his_purrr/")
}
\arguments{
\item{departamento}{nombre del departamento almacenado en cdc_peru, tipo caracter}

\item{provincia}{nombre de la provincia, tipo caracter}

\item{distrito}{nombre del distrito, tipo caracter}

\item{ruta}{especifica ruta}
}
\value{
Base de datos de morbilidad por Area administrativa
}
\description{
Llama y genera base de datos consolidada de morbilidad  segun departamento, provincia o distrito.
}
\section{Functions}{
\itemize{
\item \code{his_extraer}: Funcion para llamar base de datos por Area administrativa
}}

\examples{

# library(tidyverse)
# library(magrittr)
# library(dplyr)

# # Extraemos la base de datos
# lima<-cdc_extraer("lima")

}
