% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortalidad-data.R
\docType{data}
\name{mortdb}
\alias{mortdb}
\alias{denominadores_departamento_2000_2020_peru}
\alias{edad_estandarizada_who}
\alias{egresos_hospitalarios}
\title{CDC databases: Mortality, Hospital outflows, and more}
\format{An object of class dataset.}
\source{
CDC-MINSA, <https://www.dge.gob.pe/portal/> ; World (WHO 2000-2025) Standard <https://seer.cancer.gov/stdpopulations/world.who.html>
}
\usage{
data(mortdb)

denominadores_departamento_2000_2020_peru

edad_estandarizada_who

egresos_hospitalarios
}
\description{
Random subset of mortality (2013-2015), hospital outflows (2000-2015) databases from Peru
}
\examples{
library(cdcper)
library(tidyverse)
data(mortdb)
mortdb \%>\% glimpse()

}
\references{
CDC-MINSA
([dge.gob.pe](https://www.dge.gob.pe/portal/docs/asis/Asis_mortdb.pdf))
}
\keyword{datasets}
