% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_weight.R
\name{mutate_ewcdf}
\alias{mutate_ewcdf}
\title{Weighted ECDF}
\usage{
mutate_ewcdf(data, variable, weights)
}
\arguments{
\item{data}{input dataset}

\item{variable}{input variable}

\item{weights}{input weights}
}
\value{
tibble
}
\description{
Create tibble with weighted ECDF
}
\section{Functions}{
\itemize{
\item \code{mutate_ewcdf}: 
}}

\examples{

# source("https://raw.githubusercontent.com/NicolasWoloszko/stat_ecdf_weighted/master/stat_ecdf_weighted.R")
# https://stat.ethz.ch/pipermail/r-help/2012-October/337288.html
# https://stackoverflow.com/questions/32487457/r-ggplot-weighted-cdf

library(tidyverse)

x <- rnorm(100)
w <- runif(100)

a <- ecdf(x = x)
b <- spatstat.geom::ewcdf(x = x,weights = w)

#plot(a)
#plot(b)

tibble(x=x,w=w) \%>\%
  mutate_ewcdf(variable = x,weights = w) \%>\%
  ggplot() +
  geom_step(aes(x = x,y = ecdf_x),color="black") +
  geom_step(aes(x = x,y = ewcdf_x),color="red")
}
