% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_special.R
\name{cdc_dotwhiskers_plot}
\alias{cdc_dotwhiskers_plot}
\title{Make some special plots}
\usage{
cdc_dotwhiskers_plot(data, var_categorical, var_continuous)
}
\arguments{
\item{data}{base de datos}

\item{var_categorical}{variable categórica (idealmente politómica)}

\item{var_continuous}{variable numérica contínua}
}
\value{
estimacion puntual e incertidumbre de la distribución de una variable por cada nivel de categorias
}
\description{
Make a dot-whiskers plot (a.k.a starwars TE fighter spaceship plot!)
}
\section{Functions}{
\itemize{
\item \code{cdc_dotwhiskers_plot}: 
}}

\examples{

# not yet
# but check this out: https://twitter.com/robinson_es/status/1193204992120958976/photo/1

}
