% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_edades.R
\name{cdc_edades_peru}
\alias{cdc_edades_peru}
\alias{cdc_edades_clean}
\title{Clean age categories}
\usage{
cdc_edades_peru(data, variable_edad)

cdc_edades_clean(data, variable_edad, tipo_edad)
}
\arguments{
\item{data}{input dataset}

\item{variable_edad}{edad como variable continua}

\item{tipo_edad}{tipo de edad: A, M, D o año, mes, día}
}
\value{
tidy categorical variables from one continuous age variable
}
\description{
Categorize continuous age into classical levels
}
\section{Functions}{
\itemize{
\item \code{cdc_edades_peru}: 

\item \code{cdc_edades_clean}: priorización con dos covariables
}}

\examples{

library(tidyverse)
library(charlatan)
library(skimr)
library(rlang)
data_edad <- tibble(age=charlatan::ch_integer(n = 100,min = 2,max = 100))
data_edad \%>\% skimr::skim_without_charts()
data_edad \%>\%
  cdc_edades_peru(variable_edad = age)
data_edad \%>\%
  cdc_edades_peru(variable_edad = age) \%>\%
  select(age,edad_quinquenal) \%>\%
  group_by(edad_quinquenal) \%>\%
  skimr::skim() \%>\%
  select(edad_quinquenal,numeric.p0:numeric.p100)

}
