% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_priorities.R
\name{cdc_pareto_lista}
\alias{cdc_pareto_lista}
\alias{cdc_pareto_lista_2}
\alias{cdc_carga_coalesce}
\alias{cdc_pareto_plot}
\title{Priorizar listados en base a multiples covariables}
\usage{
cdc_pareto_lista(data, variable, pareto_cut = 85)

cdc_pareto_lista_2(data, variable_c, variable_d, pareto_cut = 80)

cdc_carga_coalesce(data)

cdc_pareto_plot(
  cdc_pareto_lista,
  pct_,
  cum_,
  variable_value,
  variable_label,
  with_format = TRUE
)
}
\arguments{
\item{data}{base de datos}

\item{variable}{variable numérica contínua bajo la cual priorizar algún listado}

\item{pareto_cut}{punto de corte tipo pareto}

\item{variable_c}{variable continua}

\item{variable_d}{variable discreta}

\item{cdc_pareto_lista}{resultado de cdc_pareto_lista}

\item{pct_}{porcentaje individual}

\item{cum_}{porcenjate acumulado}

\item{variable_value}{nombre de la variable numerica evaluada}

\item{variable_label}{nombre de la variable para las etiquetas}

\item{with_format}{add axis scale, fixed coordinates and ggrepel labels}
}
\value{
selección por criterio pareto y coalescencia para multiples covariables
}
\description{
Mostrar qué elementos representan el 80% de la suma total
}
\section{Functions}{
\itemize{
\item \code{cdc_pareto_lista}: 

\item \code{cdc_pareto_lista_2}: priorización con dos covariables

\item \code{cdc_carga_coalesce}: coalescencia de multiples covariables

\item \code{cdc_pareto_plot}: grafico de pareto: % individual vs % acumulado
}}

\examples{

\dontrun{

library(tidyverse)
library(charlatan)

n_obs <- 11
set.seed(n_obs)

ch_data_wide <- tibble(
  #names
  name = ch_currency(n = n_obs),
  #values
  category = ch_integer(n = n_obs,min = 0,max = 1) \%>\% as.logical(),
  # category_02 = ch_integer(n = n_obs,min = 0,max = 1) \%>\% as.logical(),
  value_01 = ch_beta(n = n_obs,shape1 = 2,shape2 = 8),
  value_02 = ch_integer(n = n_obs,min = 1.2,max = 9.6)) \%>\%
  pivot_longer(cols = value_01:value_02,
               names_to = "variable",
               values_to = "numeric") \%>\%
  mutate(beta = ch_beta(n = n_obs*2,shape1 = 1,shape2 = 8))

cdcper::cdc_pareto_lista(data = ch_data_wide,
                         variable = numeric,
                         pareto_cut = 80) \%>\%
  epihelper::print_inf()

#cdcper::cdc_pareto_lista_2(data = ch_data_wide,
#                          variable_c = numeric,
#                          variable_d = category,
#                          pareto_cut = 80) \%>\%
# epihelper::print_inf()
#
#cdcper::cdc_pareto_lista(data = ch_data_wide,
#                        variable = numeric,
#                        pareto_cut = 80) \%>\%
# cdcper::cdc_pareto_lista(variable = beta,
#                          pareto_cut = 85) \%>\%
# epihelper::print_inf()
#
#cdcper::cdc_pareto_lista(data = ch_data_wide,
#                        variable = numeric,
#                        pareto_cut = 80) \%>\%
# cdcper::cdc_pareto_lista(variable = beta,
#                          pareto_cut = 85) \%>\%
# cdcper::cdc_carga_coalesce() \%>\%
# epihelper::print_inf()

cdcper::cdc_pareto_lista(data = ch_data_wide,
                         variable = numeric,
                         pareto_cut = 80) \%>\%
  cdcper::cdc_pareto_plot(pct_ = pct_numeric,
                          cum_ = cum_numeric,
                          variable_value = numeric,
                          variable_label = name) #\%>\%
  # plotly::ggplotly()

}

}
