% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_met_doc.r
\name{get_data_gri}
\alias{get_data_gri}
\alias{get_ids}
\title{Descarga de información meteorológica}
\usage{
get_data_gri(var, dep = NULL, pro = NULL, dis = NULL, ids = NULL,
  fuente = NULL, ruta = NULL, rango = NULL, esc_temp = NULL,
  down = FALSE, filename = paste0("Grilla_", var, ".csv"))

get_ids(fuente = "local", ruta = NULL)
}
\arguments{
\item{var}{Variable que será buscada. Puede ser "Prec", "Tmax" o "Tmin".}

\item{dep}{Nombre o código (en caracter) del departamento.}

\item{pro}{Nombre o código (en caracter) de la provincia.}

\item{dis}{Nombre o código (en caracter) del distrito.}

\item{ids}{Objeto que contiene la metadata.}

\item{fuente}{Ubicación de la base de datos. Esta puede ser "local" o "web".}

\item{ruta}{Dirección de la carpeta que contiene los datos. Solo funciona cuando fuente = "local".}

\item{rango}{Fecha inicial y final que será filtrada.}

\item{esc_temp}{Escala temporal en la que se descargará la información. Puede ser "Semana", "Mes", "Año" o NULL (vendría a ser escala diaria).}

\item{down}{Indique si la información se descargará en un archivo (TRUE) o no (FALSE).}

\item{filename}{Nombre del archivo donde se exportará la información.}
}
\value{
Tabla con información de una variable meteorológica para una lugar, tiempo y escala temporal específica.
}
\description{
Funciones que permiten la descarga de información meteorológica (v0.2.1)
}
\details{
Usos

- La información debe ser utilizada de manera referencial y no debe ser considerado como el dato real.

- Utilizar la información para asociar el comportamiento temporal del clima con alguna variable a estudiar.

- La información debe ser utilizada como representativa del área de un distrito y no de una ubicación específica.

- La información puede ser utilizada para determinar el comportamiento espacial del clima entre distritos

- La información puede ser utilizada para analizar eventos extremos, mas no para determinar el valor numérico del evento.
}
\section{Functions}{
\itemize{
\item \code{get_data_gri}: función que realiza la descarga de información

\item \code{get_ids}: función que contiene la ubicación de la información
}}

\examples{

library(dplyr)
library(googledrive)

# Por primera vez: poner 1 -> ir a correo -> aceptar
ids.web <- get_ids(fuente = "web")

prec <- get_data_gri(
   var = "Prec",
	  dep = "LORETO",
   pro = "MAYNAS",
   dis = "IQUITOS",
   fuente = "web",
   ids = ids.web,
   esc_temp = "Mes",
   rango = c("2010-01-01","2018-12-31"))

# final product
prec

}
