% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_his.R
\name{his_create}
\alias{his_create}
\alias{his_extract}
\title{Acceder a bases del HIS MIS}
\usage{
his_create(data_index, i_info)

his_extract(data_index, i_ord)
}
\arguments{
\item{data_index}{dataframe con indices de ingreso}

\item{i_info}{indice de consulta}

\item{i_ord}{identificador}
}
\value{
Bases limpias de HIS MIS.
}
\description{
Limpiar salida y generar bases del HIS MIS.
}
\section{Functions}{
\itemize{
\item \code{his_create}: funcion para limpiar bases de HIS MIS

\item \code{his_extract}: unir bases por departamento
}}

\examples{

# ejemplo
# falta

}
