% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortdb-data.R
\docType{data}
\name{mortdb}
\alias{mortdb}
\title{Mortality database}
\format{An object of class dataset.}
\source{
CDC-MINSA, <https://www.dge.gob.pe/portal/>
}
\usage{
data(mortdb)
}
\description{
Random subset of mortality database from Peru 2013-15
}
\examples{
library(tidyverse)
data(mortdb)
mortdb \%>\% glimpse()

}
\references{
CDC-MINSA
([dge.gob.pe](https://www.dge.gob.pe/portal/docs/asis/Asis_mortalidad.pdf))
}
\keyword{datasets}
