% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_reunis.R
\name{read_reunis_total}
\alias{read_reunis_total}
\alias{read_reunis_edad}
\alias{poblacion_distrital_clean}
\alias{read_inei_poblacion}
\title{Tidy up REUNIS tables}
\usage{
read_reunis_total(file, year)

read_reunis_edad(file, year)

poblacion_distrital_clean(path)

read_inei_poblacion(path, file_name)
}
\arguments{
\item{file}{nombre de archivo}

\item{year}{anho de archivo}

\item{path}{paste path to folder of files}

\item{file_name}{name of file that corresponds to one department}
}
\value{
tidy tables a partir de poblacion a nivel nacional de la REUNIS
}
\description{
Tidy up REUNIS tables of population at district level. Ver: [REUNIS](https://www.minsa.gob.pe/reunis/data/poblacion_estimada.asp).
}
\section{Functions}{
\itemize{
\item \code{read_reunis_total}: 

\item \code{read_reunis_edad}: priorización con dos covariables

\item \code{poblacion_distrital_clean}: priorización con dos covariables

\item \code{read_inei_poblacion}: priorización con dos covariables
}}

\examples{

# read_reunis_total(file = "data-raw/Poblacion Peru 2020 Dpto Prov Dist Final INEI-actualizado.xlsx",year = 2020)

}
