% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_reunis.R
\name{read_reunis_total}
\alias{read_reunis_total}
\alias{read_reunis_edad}
\alias{poblacion_distrital_clean}
\alias{read_inei_poblacion}
\alias{read_reunis_edad_pediatricx}
\alias{read_reunis_edad_quinquenio}
\title{Tidy up REUNIS tables}
\usage{
read_reunis_total(file, year)

read_reunis_edad(file, year)

poblacion_distrital_clean(path)

read_inei_poblacion(path, file_name)

read_reunis_edad_pediatricx(file, year)

read_reunis_edad_quinquenio(file, year)
}
\arguments{
\item{file}{nombre de archivo}

\item{year}{anho de archivo}

\item{path}{paste path to folder of files}

\item{file_name}{name of file that corresponds to one department}
}
\value{
tidy tables a partir de poblacion a nivel nacional de la REUNIS
}
\description{
Tidy up REUNIS tables of population at district level. Ver: [REUNIS](https://www.minsa.gob.pe/reunis/data/poblacion_estimada.asp).
}
\section{Functions}{
\itemize{
\item \code{read_reunis_total}: 

\item \code{read_reunis_edad}: tidy tables a partir de poblacion por edades a nivel nacional de la REUNIS

\item \code{poblacion_distrital_clean}: custom function ?

\item \code{read_inei_poblacion}: tidy tables from INEI population files

\item \code{read_reunis_edad_pediatricx}: tidy tables a partir de poblacion por edades (poblacion pediatrica) a nivel nacional de la REUNIS

\item \code{read_reunis_edad_quinquenio}: tidy tables a partir de poblacion por edades (por quinquenios) a nivel nacional de la REUNIS
}}

\examples{

# rute_1 <- "data-raw/Poblacion Peru 2020 Dpto Prov Dist Final INEI-actualizado.xlsx"
# read_reunis_total(file = rute_1,year = 2020)
# read_reunis_edad(file = rute_1, year = 2020)
#
# #rute_2 <- "data-raw/PERÚ - ESTIMACIONES Y PROYECCIONES DE POBLACIÓN ( BOLETINES ESPECIALES 17,18,19,20,21,22 y 36,37)/Libro18/cuadros/"
# #read_inei_poblacion(folder = rute_2,file_name = "c1600") \%>\% count(anho)

}
